server/requirements.txt
pyaudio
numpy
openai-whisper
groq
elevenlabs
websockets
playsound==1.2.2
pydub
